Learning C, day 4

+++ Week-1/Day-5/projects_day-5.c
@@ -3,8 +3,8 @@
 #include <time.h>
 
 struct MyStruct {
-    size_t spacesAmmount;
-    size_t symbolsAmmount;
+    int spacesAmmount;
+    int symbolsAmmount;
     int errorCode;
 };
